Renamed metametrics.new.js to metametrics.js
