+++ ui/pages/asset/components/native-asset.js
@@ -14,7 +14,7 @@ import {
 import { showModal } from '../../../store/actions';
 import { DEFAULT_ROUTE } from '../../../helpers/constants/routes';
 import { getURLHostName } from '../../../helpers/utils/util';
-import { MetaMetricsContext } from '../../../contexts/metametrics.new';
+import { MetaMetricsContext } from '../../../contexts/metametrics';
 import AssetNavigation from './asset-navigation';
 import AssetOptions from './asset-options';
