+++ ui/pages/index.js
@@ -5,9 +5,9 @@ import { HashRouter } from 'react-router-dom';
 import * as Sentry from '@sentry/browser';
 import { I18nProvider, LegacyI18nProvider } from '../contexts/i18n';
 import {
-  MetaMetricsProvider as NewMetaMetricsProvider,
-  LegacyMetaMetricsProvider as NewLegacyMetaMetricsProvider,
-} from '../contexts/metametrics.new';
+  MetaMetricsProvider,
+  LegacyMetaMetricsProvider,
+} from '../contexts/metametrics';
 import ErrorPage from './error';
 import Routes from './routes';
 
@@ -41,15 +41,15 @@ class Index extends PureComponent {
     return (
       <Provider store={store}>
         <HashRouter hashType="noslash">
-          <NewMetaMetricsProvider>
-            <NewLegacyMetaMetricsProvider>
+          <MetaMetricsProvider>
+            <LegacyMetaMetricsProvider>
               <I18nProvider>
                 <LegacyI18nProvider>
                   <Routes />
                 </LegacyI18nProvider>
               </I18nProvider>
-            </NewLegacyMetaMetricsProvider>
-          </NewMetaMetricsProvider>
+            </LegacyMetaMetricsProvider>
+          </MetaMetricsProvider>
         </HashRouter>
       </Provider>
     );
