+++ ui/pages/send/send-content/send-amount-row/amount-max-button/amount-max-button.js
@@ -7,7 +7,7 @@ import {
   toggleSendMaxMode,
 } from '../../../../../ducks/send';
 import { useI18nContext } from '../../../../../hooks/useI18nContext';
-import { MetaMetricsContext } from '../../../../../contexts/metametrics.new';
+import { MetaMetricsContext } from '../../../../../contexts/metametrics';
 
 export default function AmountMaxButton() {
   const isDraftTransactionInvalid = useSelector(isSendFormInvalid);
