+++ ui/pages/swaps/awaiting-signatures/awaiting-signatures.js
@@ -32,7 +32,7 @@ import {
   DISPLAY,
 } from '../../../helpers/constants/design-system';
 import SwapsFooter from '../swaps-footer';
-import { MetaMetricsContext } from '../../../contexts/metametrics.new';
+import { MetaMetricsContext } from '../../../contexts/metametrics';
 import SwapStepIcon from './swap-step-icon';
 
 export default function AwaitingSignatures() {
