+++ ui/pages/swaps/searchable-item-list/item-list/item-list.component.js
@@ -13,7 +13,7 @@ import {
 } from '../../../../selectors';
 import { SWAPS_CHAINID_DEFAULT_BLOCK_EXPLORER_URL_MAP } from '../../../../../shared/constants/swaps';
 import { getURLHostName } from '../../../../helpers/utils/util';
-import { MetaMetricsContext } from '../../../../contexts/metametrics.new';
+import { MetaMetricsContext } from '../../../../contexts/metametrics';
 
 export default function ItemList({
   results = [],
