+++ ui/pages/swaps/smart-transaction-status/smart-transaction-status.js
@@ -44,7 +44,7 @@ import { SMART_TRANSACTION_STATUSES } from '../../../../shared/constants/transac
 import SwapsFooter from '../swaps-footer';
 import { calcTokenAmount } from '../../../helpers/utils/token-util';
 import { showRemainingTimeInMinAndSec } from '../swaps.util';
-import { MetaMetricsContext } from '../../../contexts/metametrics.new';
+import { MetaMetricsContext } from '../../../contexts/metametrics';
 import SuccessIcon from './success-icon';
 import RevertedIcon from './reverted-icon';
 import CanceledIcon from './canceled-icon';
